feat: offload ffmpeg trimming to worker threads
