+++ apps/web/package.json
@@ -54,6 +54,7 @@
     "react-use-gesture": "^9.1.3",
     "recharts": "^2.8.0",
     "tailwindcss-plugin-rtl": "^1.2.0",
+    "threads": "^1.7.0",
     "video.js": "^8.23.4",
     "web-push": "^3.6.0",
     "zod": "^4.0.16",
